refactor: lib 내 파일 -> index.js를 통해 import하는 것으로 통일
