+++ src/lib/createRouter.js
@@ -1,16 +1,16 @@
-import { createObserver } from "./createObserver";
+import { createObserver } from "./";
 
 export const createRouter = (routes) => {
-  const { subscribe, notify } = createObserver();
+	const { subscribe, notify } = createObserver();
 
-  const getTarget = () => routes[window.location.pathname]
+	const getTarget = () => routes[window.location.pathname];
 
-  const push = (path) => {
-    window.history.pushState(null, null, path);
-    notify();
-  }
+	const push = (path) => {
+		window.history.pushState(null, null, path);
+		notify();
+	};
 
-  window.addEventListener('popstate', () => notify());
+	window.addEventListener("popstate", () => notify());
 
-  return { push, subscribe, getTarget }
+	return { push, subscribe, getTarget };
 };
