+++ src/lib/createStore.js
@@ -1,16 +1,16 @@
-import { createObserver } from "./createObserver.js";
+import { createObserver } from "./";
 
 export const createStore = (initialStore) => {
-  const { subscribe, notify } = createObserver();
+	const { subscribe, notify } = createObserver();
 
-  let state = { ...initialStore };
+	let state = { ...initialStore };
 
-  const setState = (newState) => {
-    state = { ...state, ...newState }
-    notify()
-  }
+	const setState = (newState) => {
+		state = { ...state, ...newState };
+		notify();
+	};
 
-  const getState = () => ({ ...state });
+	const getState = () => ({ ...state });
 
-  return { getState, setState, subscribe }
-}
+	return { getState, setState, subscribe };
+};
