+++ src/lib/index.js
@@ -1,6 +1,9 @@
 export * from "./createStore";
 export * from "./createStorage";
+export * from "./createObserver";
 export * from "./createRouter";
 export * from "./createVNode";
 export * from "./createElement";
+export * from "./createElement__v2";
 export * from "./renderElement";
+export * from "./eventManager";
