+++ src/utils/attributesUtils.js
@@ -1,4 +1,4 @@
-import { addEvent, removeEvent } from "../lib/eventManager";
+import { addEvent, removeEvent } from "../lib";
 
 export const camelToKebab = (str) => {
 	return str.replace(/[A-Z]/g, (match) => `-${match.toLowerCase()}`);
